+++ src/discrete-layer/components/job-manager/cell-renderer/job-details/job-details.fail-reason-area.tsx
@@ -1,8 +1,7 @@
 import React,{useState, useEffect} from 'react';
-import CopyToClipboard from 'react-copy-to-clipboard';
-import { useIntl } from 'react-intl';
-import { IconButton, Tooltip, Typography } from '@map-colonies/react-core';
+import { Typography } from '@map-colonies/react-core';
 import { Box } from '@map-colonies/react-components';
+import { Copy } from '../../../../../common/components/copy/copy';
 
 import './job-details.fail-reason-area.css';
 
@@ -12,19 +11,12 @@ interface FailReasonAreaProps {
   key?: string;
 }
 
-export const FailReasonArea: React.FC<FailReasonAreaProps> = ({
-  failReason,
-  show,
-  key = '',
-}) => {
-  const intl = useIntl();
-
+export const FailReasonArea: React.FC<FailReasonAreaProps> = ({ failReason, show, key = '' }) => {
   const [containerClass, setContainerClass] = useState('failReasonAreaContainer');
 
   useEffect(() => {
     setContainerClass(`failReasonAreaContainer ${show ? 'show' : ''}`);
-  }, [show])
-
+  }, [show]);
 
   return (
     <td colSpan={5}>
@@ -32,12 +24,7 @@ export const FailReasonArea: React.FC<FailReasonAreaProps> = ({
         <Typography className="failReasonText" tag="p">
           {failReason}
         </Typography>
-
-        <Tooltip content={intl.formatMessage({ id: 'action.copy.tooltip' })}>
-          <CopyToClipboard text={failReason}>
-            <IconButton type="button" className="mc-icon-Copy" />
-          </CopyToClipboard>
-        </Tooltip>
+        <Copy value={failReason}/>
       </Box>
     </td>
   );