+++ src/discrete-layer/components/layer-details/field-value-presentors/json.value-presentor.tsx
@@ -1,16 +1,16 @@
 import React, { useState, useEffect, useRef, useMemo } from 'react';
 import { useIntl } from 'react-intl';
-import CopyToClipboard from 'react-copy-to-clipboard';
 import { get, isEmpty, set, unset } from 'lodash';
 import { Geometry } from 'geojson';
 import shp, { FeatureCollectionWithFilename, parseShp } from 'shpjs';
 import { useDebouncedCallback } from 'use-debounce';
-import { Button, IconButton, TextField, Tooltip, Typography } from '@map-colonies/react-core';
+import { Button, TextField, Typography } from '@map-colonies/react-core';
 import { Box } from '@map-colonies/react-components';
 import { EMPTY_JSON_STRING_VALUE, geoJSONValidation, validateGeoJSONString } from '../../../../common/utils/geojson.validation';
 import { emphasizeByHTML } from '../../../../common/helpers/formatters';
 import { Mode } from '../../../../common/models/mode.enum';
 import TooltippedValue from '../../../../common/components/form/tooltipped.value';
+import { Copy } from '../../../../common/components/copy/copy';
 import { geoArgs } from '../../../../common/utils/geo.tools';
 import { IRecordFieldInfo } from '../layer-details.field-info';
 import { EntityFormikHandlers } from '../layer-datails-form';
@@ -214,11 +214,7 @@ export const JsonValuePresentorComponent: React.FC<JsonValuePresentorProps> = ({
           {
             !isEmpty(value) && isCopyable &&
             <Box className="detailsFieldCopyIcon detailsFieldNoMargin">
-              <Tooltip content={intl.formatMessage({ id: 'action.copy.tooltip' })}>
-                <CopyToClipboard text={stringifiedValue}>
-                  <IconButton type="button" className="mc-icon-Copy" />
-                </CopyToClipboard>
-              </Tooltip>
+              <Copy value={stringifiedValue}/>
             </Box>
           }
         </Box>