+++ src/discrete-layer/components/layer-details/raster/job-info.tsx
@@ -5,7 +5,7 @@ import { Typography, useTheme } from '@map-colonies/react-core';
 import { Skeleton } from '../../../../common/components/skeleton/skeleton';
 import { AutoDirectionBox } from '../../../../common/components/auto-direction-box/auto-direction-box.component';
 import { Status } from '../../../models';
-import { RenderErrorCounts } from '../../job-error-summary/job-error-summary';
+import { JobErrorsSummary } from '../../job-errors-summary/job-errors-summary';
 import { Progress } from './progress';
 import { isJobValid, isStatusFailed, isTaskValid } from './state-machine/helpers';
 import { IJob } from './state-machine/types';
@@ -54,15 +54,13 @@ export const JobInfo: React.FC<JobInfoProps> = ({ job }) => {
             job.taskId ? (
               job.validationReport?.errorsSummary?.errorsCount ? (
                 <Box className="reportList bold">
-                  {
-                    RenderErrorCounts(theme, job.validationReport.errorsSummary, 'countWrapper')
-                  }
+                  {JobErrorsSummary(theme, job.validationReport.errorsSummary, "countWrapper")}
                 </Box>
               ) : (
                 <Box className="reportError error">
                   {
                     job.taskReason
-                    ? <Typography tag={'span'}>
+                    ? <Typography tag="span">
                         <AutoDirectionBox>{job.taskReason}</AutoDirectionBox>
                       </Typography>
                     : <FormattedMessage id="ingestion.error.not-found" values={{ value: 'job.validationReport.errorsSummary.errorsCount' }} />