+++ src/discrete-layer/components/layer-details/raster/state-machine/state-machine.ts
@@ -2,8 +2,8 @@ import { mergeWith } from 'lodash';
 import { assign, createMachine, fromCallback, sendParent } from 'xstate';
 import CONFIG from '../../../../../common/config';
 import { dateFormatter, relativeDateFormatter } from '../../../../../common/helpers/formatters';
+import { localStore } from '../../../../../common/helpers/storage';
 import { Mode } from '../../../../../common/models/mode.enum';
-import { Status } from '../../../../models';
 import {
   fetchProductActions,
   filesErrorActions,
@@ -291,7 +291,10 @@ export const workflowMachine = createMachine<IContext, Events>({
           target: WORKFLOW.RESTORE_JOB
         },
         RETRY: {
-          target: WORKFLOW.JOB_RETRY
+          target: WORKFLOW.RETRY_JOB
+        },
+        CLEAN_ERRORS: {
+          actions: assign({ errors: [] })
         },
         "*": { actions: warnUnexpectedStateEvent }
       }
@@ -402,12 +405,12 @@ export const workflowMachine = createMachine<IContext, Events>({
         src: SERVICES[WORKFLOW.ROOT].jobPollingService,
         onDone: [
           // {
-          //   // guard: (_: { context: IContext; event: any }) => {
-          //   //   console.log(`>>> Enter GUARD onDone of ${WORKFLOW.JOB_POLLING}`);
-          //   //   return _.event.output.details.status !== Status.InProgress &&
-          //   //     _.event.output.details.status !== Status.Pending &&
-          //   //     _.event.output.details.status !== Status.Suspended;
-          //   // },
+          //   guard: (_: { context: IContext; event: any }) => {
+          //     console.log(`>>> Enter GUARD onDone of ${WORKFLOW.JOB_POLLING}`);
+          //     return _.event.output.details.status !== Status.InProgress &&
+          //       _.event.output.details.status !== Status.Pending &&
+          //       _.event.output.details.status !== Status.Suspended;
+          //   },
           //   actions: assign((_: { context: IContext; event: any }) => ({
           //     job: {
           //       ..._.context.job,
@@ -423,7 +426,7 @@ export const workflowMachine = createMachine<IContext, Events>({
                 ..._.event.output
               }
             })),
-            target: WORKFLOW.JOB_POLLING_WAIT
+            target: WORKFLOW.WAIT.ROOT
           }
         ],
         onError: {
@@ -432,23 +435,59 @@ export const workflowMachine = createMachine<IContext, Events>({
         }
       }
     },
-    [WORKFLOW.JOB_POLLING_WAIT]: {
-      entry: assign((_: { context: IContext; event: any }) => ({
-        remainingTime: CONFIG.JOB_MANAGER.POLLING_CYCLE_INTERVAL / 1000
-      })),
-      invoke: {
-        src: fromCallback(({ sendBack }) => {
-          const interval = setInterval(() => {
-            sendBack({ type: 'TICK' });
-          }, 1000);
-          return () => clearInterval(interval);
-        })
+    [WORKFLOW.WAIT.ROOT]: {
+      id: WORKFLOW.WAIT,
+      type: 'parallel',
+      entry: assign((_: { context: IContext; event: any }) => {
+        console.log(`>>> Enter ${WORKFLOW.WAIT.ROOT}`);
+        return {
+          remainingTime: CONFIG.JOB_MANAGER.POLLING_CYCLE_INTERVAL / 1000
+        };
+      }),
+      states: {
+        [WORKFLOW.WAIT.TIMER]: {
+          invoke: {
+            src: fromCallback(({ sendBack }) => {
+              const interval = setInterval(() => {
+                sendBack({ type: 'TICK' });
+              }, 1000);
+              return () => clearInterval(interval);
+            })
+          }
+        },
+        [WORKFLOW.WAIT.WATCHER]: {
+          invoke: {
+            input: (_: { context: IContext; event: any }) => _,
+            src: fromCallback(({ sendBack, input }) => {
+              const { jobId, taskId } = input.context.job || {};
+              const getTaskNotification = () => {
+                const lastTask = localStore.get('lastTask');
+                if (!lastTask) { return; }
+                const notification = JSON.parse(lastTask);
+                if (notification && jobId && taskId && notification.jobId === jobId && notification.taskId === taskId) {
+                  sendBack({ type: 'SYNC' });
+                }
+              };
+              localStore.watchMethods(['setItem'], undefined, (_method, key) => {
+                if (key === 'MC-lastTask') {
+                  getTaskNotification();
+                }
+              });
+              return () => {
+                localStore.unWatchMethods();
+              };
+            })
+          }
+        }
       },
       on: {
         TICK: {
           actions: assign((_: { context: IContext; event: any }) => ({
             remainingTime: _.context.remainingTime ? _.context.remainingTime - 1 : 0
           }))
+        },
+        SYNC: {
+          target: WORKFLOW.JOB_POLLING
         },
         STOP_POLLING: {
           target: WORKFLOW.IDLE
@@ -477,8 +516,8 @@ export const workflowMachine = createMachine<IContext, Events>({
         }
       }
     },
-    [WORKFLOW.JOB_RETRY]: {
-      entry: () => console.log(`>>> Enter ${WORKFLOW.JOB_RETRY}`),
+    [WORKFLOW.RETRY_JOB]: {
+      entry: () => console.log(`>>> Enter ${WORKFLOW.RETRY_JOB}`),
       tags: [STATE_TAGS.GENERAL_LOADING],
       invoke: {
         input: (_: { context: IContext; event: any }) => _,