+++ src/discrete-layer/components/system-status/system-core-info/external-services/external-services.tsx
@@ -1,64 +1,51 @@
 import React from 'react';
+import { FormattedMessage } from 'react-intl';
 import { Box } from '@map-colonies/react-components';
 import {
   CollapsibleList,
-  IconButton,
   SimpleListItem,
-  Tooltip,
   Typography,
 } from '@map-colonies/react-core';
-import CopyToClipboard from 'react-copy-to-clipboard';
-import { FormattedMessage, useIntl } from 'react-intl';
+import { Copy } from '../../../../../common/components/copy/copy';
 import { CategorizedServices } from '../system-core-info.dialog';
+
 import './external-services.css';
 
 interface ExternalServicesProps {
   services: CategorizedServices;
 }
 
-export const ExternalServices: React.FC<ExternalServicesProps> = ({
-  services,
-}: ExternalServicesProps) => {
-  const intl = useIntl();
-
+export const ExternalServices: React.FC<ExternalServicesProps> = ({ services }: ExternalServicesProps) => {
   return (
     <Box className="listsContainer">
-      {Object.entries(services).map(([category, services]) => {
-        return (
-          <CollapsibleList
-            handle={
-              <SimpleListItem
-                text={category}
-                metaIcon="chevron_right"
-              />
-            }
-          >
-            {(services ?? []).map(({ url, display }) => {
-              return (
-                <Box className="externalService">
-                  <Typography className="displayName" tag="p">
-                    <FormattedMessage id={display} />
-                  </Typography>
-
-                  <Typography className="urlText" tag="p">
-                    {`${url as string}`}
-                  </Typography>
-
-                  <Tooltip
-                    content={intl.formatMessage({
-                      id: 'action.copy-url.tooltip',
-                    })}
-                  >
-                    <CopyToClipboard text={url as string}>
-                      <IconButton className="mc-icon-Copy" type="button"/>
-                    </CopyToClipboard>
-                  </Tooltip>
-                </Box>
-              );
-            })}
-          </CollapsibleList>
-        );
-      })}
+      {
+        Object.entries(services).map(([category, services]) => {
+          return (
+            <CollapsibleList
+              handle={
+                <SimpleListItem
+                  text={category}
+                  metaIcon="chevron_right"
+                />
+              }
+            >
+              {(services ?? []).map(({ url, display }) => {
+                return (
+                  <Box className="externalService">
+                    <Typography className="displayName" tag="p">
+                      <FormattedMessage id={display} />
+                    </Typography>
+                    <Typography className="urlText" tag="p">
+                      {`${url as string}`}
+                    </Typography>
+                    <Copy value={url as string}/>
+                  </Box>
+                );
+              })}
+            </CollapsibleList>
+          );
+        })
+      }
     </Box>
   );
 };